+++ libs/cdk/src/index.ts
@@ -4,10 +4,4 @@ export * from './lib/render-strategies';
 export * from './lib/template-management';
 export * from './lib/zone-less';
 export * from './lib/zone-configuration';
-export { RxStrategies } from './lib/model';
-export {
-  RxStrategyNames,
-  RxAngularConfig,
-  RxDefaultStrategyNames,
-} from './lib/model';
 export * from './lib/cdk-config';