+++ libs/cdk/src/lib/render-strategies/index.ts
@@ -1,7 +1,7 @@
-export { RxNativeStrategies, RX_NATIVE_STRATEGIES } from './native-strategies';
-export {
-  RxConcurrentStrategies,
-  RX_CONCURRENT_STRATEGIES,
-} from './concurrent-strategies';
 export { RxStrategyProvider } from './strategy-provider.service';
 export { ScheduleOnStrategyOptions } from './model';
+export {
+  RX_CONCURRENT_STRATEGIES,
+  RxConcurrentStrategies,
+} from './concurrent-strategies';
+export { RX_NATIVE_STRATEGIES, RxNativeStrategies } from './native-strategies';